// src/glance/types.ts
const enum ShaderStage
{
    VERTEX = 0x8B31,
    FRAGMENT = 0x8B30,
}


/**
 * Attachment type for the Framebuffer.
 */
const enum AttachmentType
{
    TEXTURE = 1,
    RENDERBUFFER = 2,
}
// ...
const enum TextureTarget
{
    TEXTURE_2D = 0x0DE1,
    TEXTURE_3D = 0x806F,
    TEXTURE_CUBE_MAP = 0x8513,
    TEXTURE_2D_ARRAY = 0x8C1A,
}


/**
 * All WebGL Texture data targets.
 * See https://developer.mozilla.org/en-US/docs/Web/API/WebGLRenderingContext/texImage2D#target
 * and https://developer.mozilla.org/en-US/docs/Web/API/WebGL2RenderingContext/texImage3D#target
 */
const enum TextureDataTarget
{
    TEXTURE_2D = 0x0DE1,
    TEXTURE_CUBE_MAP_POSITIVE_X = 0x8515,
    TEXTURE_CUBE_MAP_NEGATIVE_X = 0x8516,
    TEXTURE_CUBE_MAP_POSITIVE_Y = 0x8517,
    TEXTURE_CUBE_MAP_NEGATIVE_Y = 0x8518,
    TEXTURE_CUBE_MAP_POSITIVE_Z = 0x8519,
    TEXTURE_CUBE_MAP_NEGATIVE_Z = 0x851A,
    TEXTURE_3D = 0x806F,
    TEXTURE_2D_ARRAY = 0x8C1A,
}
// ...
const enum TextureInternalFormat
{
    R8 = 0x8229,
    R16F = 0x822D,
    R32F = 0x822E,
    R8UI = 0x8232,
    RG8 = 0x822B,
    RG16F = 0x822F,
    RG32F = 0x8230,
    RG8UI = 0x8238,
    RGB8 = 0x8051,
    SRGB8 = 0x8C41,
    RGB565 = 0x8D62,
    R11F_G11F_B10F = 0x8C3A,
    RGB9_E5 = 0x8C3D,
    RGB16F = 0x881B,
    RGB32F = 0x8815,
    RGB8UI = 0x8D7D,
    RGBA8 = 0x8058,
    SRGB8_ALPHA8 = 0x8C43,
    RGB5_A1 = 0x8057,
    RGBA4 = 0x8056,
    RGB10_A2 = 0x8059,
    RGBA16F = 0x881A,
    RGBA32F = 0x8814,
    RGBA8UI = 0x8D7C,
    DEPTH_COMPONENT16 = 0x81A5,
    DEPTH_COMPONENT24 = 0x81A6,
    DEPTH_COMPONENT32F = 0x8CAC,
    DEPTH24_STENCIL8 = 0x88F0,
    DEPTH32F_STENCIL8 = 0x8CAD,
    STENCIL_INDEX8 = 0x8D48,
}


/**
 * Valid filter values for the min- and magFilter parameters of texParameteri.
 * See https://developer.mozilla.org/en-US/docs/Web/API/WebGL_API/Constants#textures
 */
const enum TextureFilter
{
    NEAREST = 0x2600,
    LINEAR = 0x2601,
    NEAREST_MIPMAP_NEAREST = 0x2700,
    LINEAR_MIPMAP_NEAREST = 0x2701,
    NEAREST_MIPMAP_LINEAR = 0x2702,
    LINEAR_MIPMAP_LINEAR = 0x2703,
}
// ...
const enum TextureCompareFunc
{
    NONE = 0x0000,      // Disable depth comparison, implies TEXTURE_COMPARE_MODE = gl.NONE.
    NEVER = 0x0200,     // Always returns 0.
    LESS = 0x0201,      // Returns 1 if the new value is less than the stored value.
    EQUAL = 0x0202,     // Returns 1 if the new value is equal to the stored value.
    LEQUAL = 0x0203,    // Returns 1 if the new value not greater than the stored value.
    GREATER = 0x0204,   // Returns 1 if the new value is strictly greater than the stored value.
    NOTEQUAL = 0x0205,  // Returns 1 if the new value is not equal to the stored value.
    GEQUAL = 0x0206,    // Returns 1 if the new value is not less than the stored value.
    ALWAYS = 0x0207,    // Always returns 1.
}
// ...
type Cached<Type> = {
    /** The cached object. */
    readonly object: Type,

    /** The generation when the object was last used. */
    generation: number,
};


/**
 * A "glancified" WebGL2 context that keeps track of all created objects.
 */
type WebGL2 = WebGL2RenderingContext & {
    __glance: {
        /** The current generation of the context. */
        generation: number,

        // Some of the objects can be cached, because they can be uniquely
        // identified by some keyable value. This is useful for reusing objects
        // across runs.

        /** All loaded shaders by source. */
        shaders: Map<string, Cached<Shader>>,

        /** All linked programs by concatenated sources. */
        programs: Map<string, Cached<Program>>,

        // Other objects cannot be cached as they are not uniquely identifiable
        // by some keyable value, or at least not without significant overhead.
        // We still need to keep track of them though, to delete them later.

        /** All raw WebGL buffers. */
        buffers: Array<WebGLBuffer>,

        /** All raw WebGL VAOs. */
        vaos: Array<WebGLVertexArrayObject>,

        /** All raw WebGL textures. */
        textures: Array<WebGLTexture>,

        /** All raw WebGL renderbuffers. */
        renderbuffers: Array<WebGLRenderbuffer>,

        /** All raw WebGL framebuffers. */
        framebuffers: Array<WebGLFramebuffer>,

        // TODO: at least for the uncached objects, I could also hijack the gl.create* functions
        // to automatically add the created object to the respective array. This would be a bit
        // more robust than manually adding them to the arrays. Then again, in order to co-exist
        // with raw OpenGL Code, maybe this is actually not a good idea?
        // TODO: we could cache the buffers and textures more agressively, by hashing their data
        // and using that as the key. Ideally we'd be able to cache all the things.

    },
};


/** The location of an Attribute in a Shader Program. */
type AttributeLocation = number;


/** The ID of a WebGL Texture Unit. */
type TextureUnitId = number;


/** Any value that can be assigned to a Uniform. */
type UniformValue = number | Array<number> | Vec2 | Vec3 | Mat3 | Mat4;


/**
 * All information about an Attribute in a Shader Program.
 */
type ShaderAttribute = {
    /** The GLSL data type of the attribute. */
    readonly type: GlslAttributeDataType,

    /** The location of the attribute on the shader program. */
    readonly location: number,

    /** The precision qualifier of the attribute. Defaults to undefined. */
    readonly precision?: GlslPrecision,
};
// ...
type Uniform = {
    /** The GLSL data type of the uniform. */
    readonly type: GlslUniformDataType,

    /** The WebGL location of the uniform on the shader program. */
    readonly location: WebGLUniformLocation,

    /** The size of the uniform (for arrays, =1 if it is a scalar). */
    readonly size: number,

    /** The precision qualifier of the uniform. Defaults to undefined. */
    readonly precision?: GlslPrecision,

    /** The current value of the uniform (mutable). */
    value: UniformValue,
};


/**
 * All information about a Shader (Stage).
 */
type Shader = {
    /** The WebGL shader object. */
    readonly glo: WebGLShader;

    /** The name of the shader. */
    readonly name: string;

    /** The shader stage (vertex or fragment). */
    readonly stage: ShaderStage;

    /** The shader source code. */
    readonly source: string;
};


/** A vertex shader. */
type VertexShader = Shader & { readonly stage: ShaderStage.VERTEX; };


/** A fragment shader. */
type FragmentShader = Shader & { readonly stage: ShaderStage.FRAGMENT; };


/**
 * All information about a Shader Program.
 */
type Program = {
    /** The WebGL program object. */
    readonly glo: WebGLProgram;

    /** The name of the program. */
    readonly name: string;

    /** The vertex shader of the program. */
    readonly vertex: VertexShader;

    /** The fragment shader of the program. */
    readonly fragment: FragmentShader;

    /** All attributes of the program. */
    readonly attributes: Map<string, ShaderAttribute>;

    /** All uniforms of the program. */
    readonly uniforms: Map<string, Uniform>;
};
// ...
type Texture = {
    /** The WebGL texture object. */
    readonly glo: WebGLTexture,

    /** The name of the Texture. */
    readonly name: string,

    /** The target of the texture. */
    readonly target: TextureTarget,

    /** The width of the texture. In Cubemaps this is the size of one face. */
    readonly width: number,

    /** The height of the texture. In Cubemaps this is equal to the height. */
    readonly height: number,

    /** The depth of the texture. 2D Textures have a depth of 1. */
    readonly depth: number,

    /** The number of mipmap levels. */
    readonly levels: number,

    /** The internal format of the texture. */
    readonly internalFormat: TextureInternalFormat;

    /** Texture comparison format. Default is TextureCompareFunc.NONE. */
    readonly compareFunc: TextureCompareFunc;

    /** The Renderbuffer attachment type. */
    readonly attachmentType: AttachmentType.TEXTURE,
};
// ...
type FramebufferAttachment = Texture | Renderbuffer;


/**
 * Description of an attachment for the Framebuffer.
 * Has a lof of optional fields, because it is user-defined.
 */
type AttachmentDefinition = {
    /** The attachment. */
    readonly attachment: FramebufferAttachment,

    /** Which mip-map to attach. Defaults to zero. */
    readonly level?: number,

    /** Attachment target (only useful for cube maps). */
    readonly target?: TextureDataTarget,

    /** Which layer to attach (only useful for 3D and 2D Array textures). */
    readonly layer?: number,
};


/**
 * A WebGL Framebuffer.
 */
type Framebuffer = {
    /** The WebGL framebuffer. */
    readonly glo: WebGLFramebuffer,

    /** The name of the framebuffer */
    readonly name: string,

    /** Color attachements. */
    readonly color: Array<AttachmentDefinition>,

    /** Depth attachment. */
    readonly depth: AttachmentDefinition | null,

    /** Stencil attachment. */
    readonly stencil: AttachmentDefinition | null,
};